Moved apio_console.py to a shared directory apio.common. This file will be used by both the parent apio process and the child scons process.
