+++ apio/commands/apio_raw.py
@@ -11,7 +11,7 @@ import sys
 import subprocess
 from typing import Tuple, List
 import click
-from apio.utils.apio_console import cout
+from apio.common.apio_console import cout
 from apio.apio_context import ApioContext, ApioContextScope
 from apio.commands import options
 from apio.utils import cmd_util, pkg_util
