+++ apio/commands/apio_sim.py
@@ -10,7 +10,7 @@
 import sys
 from pathlib import Path
 import click
-from apio.utils.apio_console import cout
+from apio.common.apio_console import cout
 from apio.managers.scons import SCons
 from apio.commands import options
 from apio.apio_context import ApioContext, ApioContextScope
