+++ apio/common/README.md
@@ -1,0 +1,2 @@
+The files in this directory are used in both the apio parent process
+and the scons child process.
