+++ apio/managers/examples.py
@@ -12,7 +12,7 @@ import os
 from pathlib import Path, PosixPath
 from dataclasses import dataclass
 from typing import Optional, List, Dict
-from apio.utils.apio_console import cout, cstyle, cerror
+from apio.common.apio_console import cout, cstyle, cerror
 from apio.apio_context import ApioContext
 from apio.managers import installer
