+++ apio/managers/programmers.py
@@ -11,7 +11,7 @@
 
 import re
 import sys
-from apio.utils.apio_console import cout, cerror
+from apio.common.apio_console import cout, cerror
 from apio.utils import util, pkg_util
 from apio.managers.system import System
 from apio.apio_context import ApioContext
