+++ apio/managers/system.py
@@ -8,7 +8,7 @@
 
 import re
 import sys
-from apio.utils.apio_console import cout, cerror
+from apio.common.apio_console import cout, cerror
 from apio.utils import util, pkg_util
 from apio.apio_context import ApioContext
 from apio.managers import installer
