+++ apio/scons/plugin_base.py
@@ -14,9 +14,9 @@ from dataclasses import dataclass
 from SCons.Builder import BuilderBase
 from SCons.Action import Action
 from SCons.Script import Builder
+from apio.common.apio_console import cout
 from apio.scons.apio_env import ApioEnv, TARGET, BUILD_DIR_SEP
 from apio.proto.apio_pb2 import GraphOutputType
-from apio.scons.scons_console import cprint
 from apio.scons.plugin_util import (
     SRC_SUFFIXES,
     verilog_src_scanner,
@@ -160,7 +160,7 @@ class PluginBase:
         def completion_action(source, target, env):  # noqa
             """Action function that prints a completion message."""
             _ = (source, target, env)  # Unused
-            cprint(f"Generated {TARGET}.{type_str}", style="green")
+            cout(f"Generated {TARGET}.{type_str}", style="green")
 
         actions = [
             f"dot -T{type_str} $SOURCES -o $TARGET",
