+++ apio/scons/plugin_util.py
@@ -28,7 +28,7 @@ from SCons.Node import NodeList
 from SCons.Node.Alias import Alias
 import debugpy
 from apio.scons.apio_env import ApioEnv, TARGET, BUILD_DIR_SEP
-from apio.scons.scons_console import cprint, cerror, cwarning
+from apio.common.apio_console import cout, cerror, cwarning
 
 # -- A list with the file extensions of the verilog source files.
 SRC_SUFFIXES = [".v", ".sv"]
@@ -42,11 +42,11 @@ def maybe_wait_for_remote_debugger(env_var_name: str):
     debugger (e.g. from Visual Studio Code) is attached.
     """
     if os.getenv(env_var_name) is not None:
-        cprint(f"Env var '{env_var_name}' was detected.")
+        cout(f"Env var '{env_var_name}' was detected.")
         port = 5678
-        cprint(f"Apio SCons for remote debugger on port localhost:{port}.")
+        cout(f"Apio SCons for remote debugger on port localhost:{port}.")
         debugpy.listen(port)
-        cprint(
+        cout(
             "Attach Visual Studio Code python remote python debugger "
             f"to port {port}.",
             style="magenta",
@@ -54,7 +54,7 @@ def maybe_wait_for_remote_debugger(env_var_name: str):
         # -- Block until the debugger connetcs.
         debugpy.wait_for_client()
         # -- Here the remote debugger is attached and the program continues.
-        cprint(
+        cout(
             "Remote debugger is attached, program continues...",
             style="green",
         )
@@ -201,7 +201,7 @@ def verilog_src_scanner(apio_env: ApioEnv) -> Scanner.Base:
             if Path(dependency).exists():
                 dependencies.append(dependency)
             elif apio_env.is_debug:
-                cprint(
+                cout(
                     f"Dependency candidate {dependency} does not exist, "
                     "dropping."
                 )
@@ -211,9 +211,9 @@ def verilog_src_scanner(apio_env: ApioEnv) -> Scanner.Base:
 
         # Debug info.
         if apio_env.is_debug:
-            cprint(f"Dependencies of {file_node.name}:", style="blue")
+            cout(f"Dependencies of {file_node.name}:", style="blue")
             for dependency in dependencies:
-                cprint(f"  {dependency}", style="blue")
+                cout(f"  {dependency}", style="blue")
 
         # All done
         return apio_env.scons_env.File(dependencies)
@@ -328,7 +328,7 @@ def check_valid_testbench_name(testbench: str) -> None:
     and exit."""
     if not is_verilog_src(testbench) or not has_testbench_name(testbench):
         cerror(f"'{testbench}' is not a valid testbench file name.")
-        cprint(TESTBENCH_HINT, style="yellow")
+        cout(TESTBENCH_HINT, style="yellow")
         sys.exit(1)
 
 
@@ -352,19 +352,19 @@ def get_sim_config(
         # -- Case 2 Testbench name was not specified and no testbench files
         # -- were found in the project.
         cerror("No testbench files found in the project.")
-        cprint(TESTBENCH_HINT, style="yellow")
+        cout(TESTBENCH_HINT, style="yellow")
 
         sys.exit(1)
     elif len(test_srcs) == 1:
         # -- Case 3 Testbench name was not specified but there is exactly
         # -- one in the project.
         testbench = test_srcs[0]
-        cprint(f"Found testbench file {testbench}", style="cyan bold")
+        cout(f"Found testbench file {testbench}", style="cyan")
     else:
         # -- Case 4 Testbench name was not specified and there are multiple
         # -- testbench files in the project.
         cerror("Multiple testbench files found in the project.")
-        cprint(
+        cout(
             "Please specify the testbench file name in the command "
             "or in apio.ini 'default-testbench' option.",
             style="yellow",
@@ -405,7 +405,7 @@ def get_tests_configs(
         # -- Case 2 - Testbench file name was not specified and there are no
         # -- testbench files in the project.
         cerror("No testbench files found in the project.")
-        cprint(TESTBENCH_HINT, style="yellow")
+        cout(TESTBENCH_HINT, style="yellow")
         sys.exit(1)
     else:
         # -- Case 3 - Testbench file name was not specified but there are one
@@ -470,7 +470,7 @@ def source_file_issue_action() -> FunctionAction:
                 continue
 
             # -- Here the file is a testbench file.
-            cprint(f"Testbench {file.name}", style="cyan bold")
+            cout(f"Testbench {file.name}", style="cyan")
 
             # -- Read the testbench file text.
             file_text = file.get_text_contents()
@@ -522,15 +522,15 @@ def _print_pnr_report(
     report: Dict[str, any] = json.loads(json_txt)
 
     # --- Report utilization
-    cprint("")
-    cprint("UTILIZATION:", style="cyan bold")
+    cout("")
+    cout("UTILIZATION:", style="cyan")
     utilization = report["utilization"]
     for resource, vals in utilization.items():
         available = vals["available"]
         used = vals["used"]
         percents = int(100 * used / available)
         style = "magenta" if used > 0 else None
-        cprint(
+        cout(
             f"{resource:>20}: {used:5} {available:5} {percents:5}%",
             style=style,
         )
@@ -540,8 +540,8 @@ def _print_pnr_report(
     # -- NOTE: As of Oct 2024, some projects do not generate timing
     # -- information and this is being investigated.
     # -- See https://github.com/FPGAwars/icestudio/issues/774 for details.
-    cprint("")
-    cprint("CLOCKS:", style="cyan bold")
+    cout("")
+    cout("CLOCKS:", style="cyan")
     clocks = report["fmax"]
     if len(clocks) > 0:
         for clk_net, vals in clocks.items():
@@ -550,16 +550,16 @@ def _print_pnr_report(
 
             # -- Report speed.
             max_mhz = vals["achieved"]
-            cprint(
+            cout(
                 f"{clk_signal:>20}: "
                 f"[magenta]{max_mhz:7.2f}[/magenta] Mhz max"
             )
 
     # -- For now we ignore the critical path report in the pnr report and
     # -- refer the user to the pnr verbose output.
-    cprint("")
+    cout("")
     if not verbose:
-        cprint(
+        cout(
             "Use 'apio report --verbose' for more details.",
             style="yellow",
         )
